barely there i change sommething and i have a segfault now don't now why

+++ src/utilities.c
@@ -6,17 +6,17 @@
 /*   By: aboulest <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 17:35:07 by aboulest          #+#    #+#             */
-/*   Updated: 2023/04/12 17:48:28 by aboulest         ###   ########.fr       */
+/*   Updated: 2023/04/14 12:07:10 by aboulest         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long	get_time(void)
+long long	get_time(void)
 {
-	long			time;
-	struct timeval	tv;
-	
+	long long			time;
+	struct timeval		tv;
+
 	if (gettimeofday(&tv, NULL) != 0)
 		return (-1);
 	time = tv.tv_sec * 1000 + tv.tv_usec / 1000;
@@ -25,12 +25,11 @@ long	get_time(void)
 
 void	waiting(unsigned int wait)
 {
-	long	time;
-	long	target;
+	long long	time;
+	long long	target;
 
 	time = get_time();
-	target = wait; 
-
+	target = wait;
 	while (get_time() - time < target)
 		usleep(10);
 }
